Results: Begin remove redundancy in DataContainer
Three constructors for DataContainer handling
ROI, mask and no mask or RIO
contain lots of copy paste, making it hard to
get all the places needed to change one thing.

So begin task of factoring out the redunant lines
into their own method: populateData()
to be run inside DataContainer class instances.

This version only handles the DataContainer
contructor for when there is no ROI/mask.
But it works for that case.

Signed-off-by: daniel j. white <dan@chalkie.org.uk>