Results: Broken! Refactor rest of redunadant DataContainer constructors
put the duplicated code in the second two class constructors
into the new populateData() method.

Still works for no ROI/mask, but there is a bug:

Using a ROI is not working... it reports maskType.NONE
instead of regular or irregular, and the ROI is not used
for the analysis.

Using a mask image might also be broken.

Something broke about how the masks are being setup and used.

Signed-off-by: daniel j. white <dan@chalkie.org.uk>